requirements.txt
shioaji
pandas
pandas_ta
gspread
google-auth
line-bot-sdk
python-dotenv
